Skip test suite on `aarch64` & `ppc64le`
The `ppc64le` test suite hangs on Travis (and has done so for a few
releases now). Perhaps it is just too much for that CI provider to
handle.

The `aarch64` builds need patches to LLVM to fix some test cases. Those
patches are currently in review at that feedstock. So go ahead and skip
these tests for now. Can re-enable once patched LLVM builds are
released.